requirements.txt
torch>=2.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
absl-py>=1.4.0
